requirements.txt
paho-mqtt==2.1.0
kafka-python==2.3.0
streamlit==1.55.0
pandas==2.3.3
matplotlib==3.10.8
apache-airflow==2.9.3
